requirements.txt

# Core dependencies
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0

# ML Models
xgboost==2.0.0
lightgbm==4.0.0
catboost==1.2.2

# Backend
fastapi==0.104.0
uvicorn==0.23.2
pydantic==2.3.0
pydantic-settings==2.0.3
python-multipart==0.0.6
python-dotenv==1.0.0

# Frontend
streamlit==1.28.0
plotly==5.17.0
requests==2.31.0

# Data & Features
scipy==1.11.3

# ML Tracking & Versioning
mlflow==2.7.1
dvc==3.32.0
dvc-s3==3.0.0

# Explainability
shap==0.43.0
lime==0.2.0

# Data Quality & Monitoring
evidently==0.4.11

# Utilities
joblib==1.3.2
pyyaml==6.0.1
python-dateutil==2.8.2

# Database
sqlalchemy==2.0.22
psycopg2-binary==2.9.9
alembic==1.12.1

# Testing
pytest==7.4.2
pytest-cov==4.1.0
pytest-asyncio==0.21.1
httpx==0.25.0

# Code Quality
black==23.10.1
flake8==6.1.0
pylint==3.0.2
isort==5.12.0
mypy==1.6.1

# Logging & Monitoring
python-json-logger==2.0.7
prometheus-client==0.18.0

# Development
jupyter==1.0.0
notebook==7.0.5
ipython==8.16.0

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Security
cryptography==41.0.7
